actualizacion

+++ src/Numero.java
@@ -1,21 +1,23 @@
 public class Numero {
     //Atributos de la clase Numero
-    private int numero;
+    private double numero;
     //Contructor de numero sin parametros
     public Numero() {
 
     }
 
     //Constructor de la clase numero con parametros
-    public Numero(int numero) {
+    public Numero(double numero) {
         this.numero = numero;
     }
+
     //Metodo get de el atributo nunero
-    public int getNumero() {
+    public double getNumero() {
         return numero;
     }
     //Metodo set de el atributo nunero
-    public void setNumero(int numero) {
+    public void setNumero(double numero) {
         this.numero = numero;
     }
+    
 }
